+++ src/frontend/Orion/Administrator/ListsConfigManagement/ListsPanel/Buttons/index.js
@@ -1,4 +1,5 @@
 import CreateButton from './CreateButton'
 import UpdateButton from './UpdateButton'
+import DeleteButton from './DeleteButton'
 
-export { CreateButton, UpdateButton }
+export { CreateButton, UpdateButton, DeleteButton }