+++ src/frontend/Orion/Administrator/ListsConfigManagement/ListsPanel/index.js
@@ -1,1 +1,4 @@
-export { default } from './ListsPanel'
+import ListsPanelListHeader from './ListsPanelListHeader'
+import ListsPanelListItem from './ListsPanelListItem'
+
+export { ListsPanelListHeader, ListsPanelListItem }