+++ CHANGELOG.md
@@ -12,7 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
-## [Unreleased] - Started: 4-14-21
+## [Unreleased] - Started: 5-4-21
+
+## [4.9.0] - Started: 4-14-21
+
+### Added
+- `Structural Components`: Built out structural list panels [#987](https://github.com/pulse-analytics/wave-db-mgmt/pull/987)
+- `Value Perception Tool`: Initial market basket index and detail page [#982](https://github.com/pulse-analytics/wave-db-mgmt/pull/982) + [#985](https://github.com/pulse-analytics/wave-db-mgmt/pull/985)
+
+### Changed
+- `Lists Config Mgmt`: Refactored to use structural list panels [#988](https://github.com/pulse-analytics/wave-db-mgmt/pull/988)
+
+### Fixed
+- `Vega Ops`: Fixed breakage in products, indications, regimens CRUD [#984](https://github.com/pulse-analytics/wave-db-mgmt/pull/984) + [#996](https://github.com/pulse-analytics/wave-db-mgmt/pull/996)
 
 ## [4.8.0] - Started: 3-29-21